+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased] - yyyy-mm-dd
 
 ### Added
-- Nothing
+- Added DtTileFlags ([@ikpil](https://github.com/ikpil))
 
 ### Fixed
 - Nothing